Restore checkboxes being selected on load
This reverts commit a0794395c0441a4714f8ced6c6b12728dd9f8627.

The change was made upstream to:
"change default behaviour of tag checkboxes: none checked instead of all checked"

However, this fork already de-selected most checkboxes in favor of the usec one;
this change breaks that behavior.

(It also breaks the local storage of the selected checkboxes, which seems undesirable?)

This commit restores the old behavior.